# lib/api-zod/src/generated/types/adminKpi.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminKpi {
  totalTenants: number;
  activeTenants: number;
  trialTenants: number;
  suspendedTenants: number;
  stripeConnectedTenants: number;
  stripeConfigured: boolean;
}