# lib/api-zod/src/generated/types/adminTenant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminTenantPlanTier } from "./adminTenantPlanTier";
import type { AdminTenantStatus } from "./adminTenantStatus";

export interface AdminTenant {
  id: number;
  name: string;
  slug: string;
  status: AdminTenantStatus;
  planTier: AdminTenantPlanTier;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  /** @nullable */
  onboardingCompletedAt?: string | null;
  createdAt: string;
  memberCount: number;
}