# lib/api-zod/src/generated/types/adminTenantDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminTenantDetailPlanTier } from "./adminTenantDetailPlanTier";
import type { AdminTenantDetailStatus } from "./adminTenantDetailStatus";

export interface AdminTenantDetail {
  id: number;
  name: string;
  /** @nullable */
  tradingName?: string | null;
  slug: string;
  status: AdminTenantDetailStatus;
  planTier: AdminTenantDetailPlanTier;
  /** @nullable */
  legalName?: string | null;
  /** @nullable */
  taxId?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  addressLine1?: string | null;
  /** @nullable */
  addressLine2?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  fiscalYearStart?: number | null;
  /** @nullable */
  industryType?: string | null;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  /** @nullable */
  onboardingCompletedAt?: string | null;
  createdAt: string;
  updatedAt: string;
  memberCount: number;
}