# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: number;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  actorClerkId?: string | null;
  /** @nullable */
  actorEmail?: string | null;
  action: string;
  /** @nullable */
  entityType?: string | null;
  /** @nullable */
  entityId?: string | null;
  oldValues?: unknown | null;
  newValues?: unknown | null;
  createdAt: string;
}