# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Invoice {
  id: string;
  /** @nullable */
  number?: string | null;
  /** @nullable */
  status?: string | null;
  amountDue: number;
  amountPaid: number;
  currency: string;
  created: string;
  /** @nullable */
  hostedInvoiceUrl?: string | null;
}