# lib/api-zod/src/generated/types/invoiceList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Invoice } from "./invoice";

export interface InvoiceList {
  stripeConfigured: boolean;
  invoices: Invoice[];
}